feat: all

+++ src/main/java/dev/shteryu/star_wars/repository/PeopleRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import dev.shteryu.star_wars.models.People;
 
 @Repository
-public interface PeopleRepository extends CrudRepository<People, Long> {
+public interface PeopleRepository extends CrudRepository<People, Integer> {
     
 }
